Removed LoadDatabase configuration bean; refactored into SecurityConfiguration bean
